# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nova Kudos",
        "description": "Demo Nova Kudos Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/app/events/companies/{company}": {
            "post": {
                "tags": [
                    "events"
                ],
                "summary": "store new event",
                "description": "store new event",
                "operationId": "store new event",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "price"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "yasin"
                                    },
                                    "description": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "long text"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "23"
                                    },
                                    "avatar": {
                                        "type": "file",
                                        "format": "text"
                                    },
                                    "expiration_date": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "23/12/22"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/api/app/events/companies/{company}"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "validation error"
                    },
                    "422": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/app/events/{event}/update": {
            "put": {
                "tags": [
                    "events"
                ],
                "summary": "update event",
                "description": "update event",
                "operationId": "update event",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "yasin"
                                    },
                                    "description": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "long text"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "23"
                                    },
                                    "avatar": {
                                        "type": "file",
                                        "format": "text"
                                    },
                                    "expiration_date": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "23/12/22"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/api/app/events/{event}/update"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/app/events/{event}/users/{user}": {
            "post": {
                "tags": [
                    "events"
                ],
                "summary": "add user to view event",
                "description": "add user to view eventt",
                "operationId": "add user to view event",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/api/app/events/{event}/users/{user}"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "validation error"
                    },
                    "422": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/app/events/{event}/groups/{group}": {
            "post": {
                "tags": [
                    "events"
                ],
                "summary": "add group to view event",
                "description": "add group to view eventt",
                "operationId": "add group to view event",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/api/app/events/{event}/groups/{group}"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "validation error"
                    },
                    "422": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}